tf df ready, must fix search output

+++ ergasia1/trie.h
@@ -42,6 +42,6 @@ void destroySingleNode(Trie* trie);
 arrayWords* getAllWordsOfTrie(Trie* trie);
 char* recGetWordsFromTrie(Trie* originalTrie, Trie* trie, char* word, arrayWords* array);
 
-double getScoreWithoutSum(Trie* trie, Map* map, double idf, char* word, int textId, int avgdl); 
+double getScoreWithoutSum(Trie* trie, Map* map, double idf, char* word, int textId, double avgdl); 
 
 #endif
